[TASK] Fix phpstan checkFunctionArgumentTypes errors in ext:backend Controller
This patch fixes incompatible type usage in function arguments
and is preparatory work for introducing native type hints and
strict mode in all core files.

Releases: master, 10.4
Resolves: #92578
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/66157
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>

+++ typo3/sysext/backend/Classes/Controller/NewRecordController.php
@@ -523,6 +523,7 @@ class NewRecordController
             if (is_array($GLOBALS['TCA'])) {
                 $groupName = '';
                 foreach ($GLOBALS['TCA'] as $table => $v) {
+                    /** @var string $table */
                     $rootLevelConfiguration = isset($v['ctrl']['rootLevel']) ? (int)$v['ctrl']['rootLevel'] : 0;
                     if ($table !== 'pages'
                         && $this->isRecordCreationAllowedForTable($table)
@@ -576,7 +577,7 @@ class NewRecordController
                                 $title = (string)($v['ctrl']['title'] ?? '');
                                 if (strpos($title, 'LLL:EXT:') === 0) {
                                     $_EXTKEY = substr($title, 8);
-                                    $_EXTKEY = substr($_EXTKEY, 0, strpos($_EXTKEY, '/'));
+                                    $_EXTKEY = substr($_EXTKEY, 0, (int)strpos($_EXTKEY, '/'));
                                     if ($_EXTKEY !== '') {
                                         // First try to get localisation of extension title
                                         $temp = explode(':', substr($title, 9 + strlen($_EXTKEY)));
